@@ -1,9 +1,15 @@
-<script>
+<script lang='ts'>
 	import { SKIP_TO_CONTENT_ID } from '$const'
 
-	export let id = SKIP_TO_CONTENT_ID
+	interface Props {
+		id?: string
+		children?: import('svelte').Snippet
+		[key: string]: unknown
+	}
+
+	let { id = SKIP_TO_CONTENT_ID, children, ...rest }: Props = $props()
 </script>
 
-<main {...$$restProps} {id}>
-	<slot />
+<main {...rest} {id}>
+	{@render children?.()}
 </main>
