@@ -1,13 +1,19 @@
 <script lang='ts'>
 	import { SKIP_TO_CONTENT_ID } from '$const'
 
-	export let id = SKIP_TO_CONTENT_ID
+	interface Props {
+		id?: string
+		text?: import('svelte').Snippet
+		[key: string]: unknown
+	}
+
+	let { id = SKIP_TO_CONTENT_ID, text, ...rest }: Props = $props()
 </script>
 
-<a {...$$restProps} class='skip-to-content-nav' href={`#${id}`}>
-	<slot name='text'>
+<a {...rest} class='skip-to-content-nav' href={`#${id}`}>
+	{#if text}{@render text()}{:else}
 		Skip to content
-	</slot>
+	{/if}
 </a>
 
 <style>
