@@ -1,12 +1,12 @@
 <script lang='ts'>
 	import { signIn } from '@auth/sveltekit/client'
-	import { pa } from '$lib/store/plausible'
-	import Spacer from '$lib/Spacer.svelte'
+	import { page } from '$app/state'
+	import { GITHUB_REPO_URL, TITLE } from '$const'
 	import { classList } from '$lib/actions'
 	import Primary from '$lib/button/Primary.svelte'
-	import { GITHUB_REPO_URL, TITLE } from '$const'
+	import Spacer from '$lib/Spacer.svelte'
+	import { pa } from '$lib/store/plausible'
 	import Svg from '$lib/Svg.svelte'
-	import { page } from '$app/stores'
 </script>
 
 <h1 class='visually-hidden'>{TITLE}</h1>
@@ -19,12 +19,12 @@
 		<Spacer axis='vertical' size='2xs' />
 		<p class='text-md'>Display your watched movies and shows in a poster grid. Easily switch between years and get an overview of all your history.</p>
 		<Spacer axis='vertical' size='m' />
-		{#if $page.data?.session?.user}
+		{#if page.data?.session?.user}
 			<Primary type='link' href='/dashboard'>
 				Show me my Poster Grid <Svg id='arrow-right' />
 			</Primary>
 		{:else}
-			<Primary type='text' on:click={() => {
+			<Primary type='text' onclick={() => {
 				pa.addEvent('login', { props: { position: 'hero' } })
 				signIn('trakt', { callbackUrl: '/dashboard' })
 			}}>
