@@ -1,5 +1,5 @@
-const postCssPresetEnv = require('postcss-preset-env')
 const postCssGlobalData = require('@csstools/postcss-global-data')
+const postCssPresetEnv = require('postcss-preset-env')
 
 const config = {
 	plugins: [
